frontend: Use a confirm dialog before deleting an item
Deleting an object/item is an important action that requires user
confirmation. Thus, this patch adds a new component (ConfirmDialog)
that will can be used for the confirmation.

+++ frontend/src/components/common/DeleteButton.js
@@ -5,21 +5,32 @@ import Tooltip from '@material-ui/core/Tooltip';
 import React from 'react';
 import { useDispatch } from 'react-redux';
 import { deleteClusterObjects } from '../../redux/actions/actions';
+import { ConfirmDialog } from './Dialog';
 
 export default function DeleteButton(props) {
   const dispatch = useDispatch();
   const { items, deletionCallback, options } = props;
+  const [openAlert, setOpenAlert] = React.useState(false);
 
   return (
-    <Tooltip title="Delete">
-      <IconButton
-        aria-label="delete"
-        onClick={() => {
+    <React.Fragment>
+      <Tooltip title="Delete">
+        <IconButton
+          aria-label="delete"
+          onClick={() => setOpenAlert(true)}
+        >
+          <Icon icon={deleteIcon} />
+        </IconButton>
+      </Tooltip>
+      <ConfirmDialog
+        open={openAlert}
+        title="Delete item"
+        description="Are you sure you want to delete this item?"
+        handleClose={() => setOpenAlert(false)}
+        onConfirm={() => {
           dispatch(deleteClusterObjects(items, deletionCallback, options));
         }}
-      >
-        <Icon icon={deleteIcon} />
-      </IconButton>
-    </Tooltip>
+      />
+    </React.Fragment>
   );
 }
